Move functions to correct file and name correctly
