+++ alpm.go
@@ -12,18 +12,44 @@ import "C"
 
 import "unsafe"
 
+// Init initializes alpm handle
+func Initialize(root, dbpath string) (*Handle, error) {
+	cRoot := C.CString(root)
+	cDBPath := C.CString(dbpath)
+	var cErr C.alpm_errno_t
+	h := C.alpm_initialize(cRoot, cDBPath, &cErr)
+
+	defer C.free(unsafe.Pointer(cRoot))
+	defer C.free(unsafe.Pointer(cDBPath))
+
+	if cErr != 0 {
+		return nil, Error(cErr)
+	}
+
+	return &Handle{h}, nil
+}
+
+// Release releases the alpm handle
+func (h *Handle) Release() error {
+	if er := C.alpm_release(h.ptr); er != 0 {
+		return Error(er)
+	}
+	h.ptr = nil
+	return nil
+}
+
+// LastError gets the last pm_error
+func (h *Handle) LastError() error {
+	if h.ptr != nil {
+		cErr := C.alpm_errno(h.ptr)
+		if cErr != 0 {
+			return Error(cErr)
+		}
+	}
+	return nil
+}
+
 // Version returns libalpm version string.
 func Version() string {
 	return C.GoString(C.alpm_version())
-}
-
-// VerCmp performs version comparison according to Pacman conventions. Return
-// value is <0 if and only if v1 is older than v2.
-func VerCmp(v1, v2 string) int {
-	c1 := C.CString(v1)
-	c2 := C.CString(v2)
-	defer C.free(unsafe.Pointer(c1))
-	defer C.free(unsafe.Pointer(c2))
-	result := C.alpm_pkg_vercmp(c1, c2)
-	return int(result)
 }
